better model generator command !

+++ api/new-model.sh
@@ -1,18 +1,42 @@
+#!/bin/bash
+
 FILEPATH=src/mongodb/models
 cd $FILEPATH
 
-MODEL=$1
-MODEL_LOWERCASE=$(/bin/sed -e 's/\(.*\)/\L\1/' <<<$1)
-FILE="$MODEL_LOWERCASE.model.ts"
-COLLECTION="${MODEL_LOWERCASE}s"
-if [ ! -z "$2" ]; then
-  COLLECTION=$2
-fi
+DRY_RUN=false
+SUPPLEMENTARY=()
+while [[ $# -gt 0 ]]; do
+  case "$1" in
+  --dry-run)
+    DRY_RUN=true
+    shift
+    ;;
+  --)
+    shift
+    ;;
+  *)
+    SUPPLEMENTARY+=("$1")
+    shift
+    ;;
+  esac
+done
+set -- ${SUPPLEMENTARY[@]} # restore positional parameters
 
-if [ -e $FILE ]; then
-  echo "File '$FILE' already exists at $FILEPATH/$FILE !"
-else
+NAME=$1
+KEBABCASE=$(sed --expression 's/\([A-Z]\)/-\L\1/g' --expression 's/^-//' <<<$NAME)
+CAMELCASE=$(/bin/sed -e 's/^./\L&\E/' <<<$NAME)
+FILE="$KEBABCASE.model.ts"
+COLLECTION="${CAMELCASE}s"
+
+if [ $DRY_RUN = true ]; then
+  echo "Running in 'DRY' mode (no changes)"
   echo "Generating the model at $FILEPATH/$FILE"
-  export MODEL COLLECTION
-  envsubst <"template/template" >"$FILE"
+else
+  if [ -e $FILE ]; then
+    echo "File '$FILE' already exists at $FILEPATH/$FILE !"
+  else
+    echo "Generating the model at $FILEPATH/$FILE"
+    export NAME COLLECTION
+    envsubst <"template/template" >"$FILE"
+  fi
 fi
